+++ exercises/templates/05_WinIdeaWorkspaceTemplate/src/display.h
@@ -81,9 +81,17 @@ void writeTextln(const char *text);
 void writeTextln_s(const char *text);
 
 /**
- *    Write a 16 bit variable on the display.
+ * Writes a 16 bit variable on the display.
  */
-void writeNumberOnDisplay(const uint16_t *value);
-void writeNumberOnDisplay_s(const uint16_t *value);
+void writeNumberOnDisplay(const uint8_t *value);
+void writeNumberOnDisplay_s(const uint8_t *value);
+
+/**
+ * Writes a 16 bit variable on the display (right-aligned).
+ * This function writes out six right-aligned digits ("65535\0").
+ * 
+ */
+void writeNumberOnDisplayRight(const uint8_t *value);
+void writeNumberOnDisplayRight_s(const uint8_t *value);
 
 #endif